+++ cmd/operator-sdk/alpha/cmd.go
@@ -15,8 +15,10 @@
 package alpha
 
 import (
+	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/cleanup"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/kubebuilder"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/olm"
+	run "github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/run"
 
 	"github.com/spf13/cobra"
 )
@@ -30,6 +32,8 @@ func NewCmd() *cobra.Command {
 	cmd.AddCommand(
 		olm.NewCmd(),
 		kubebuilder.NewCmd(),
+		run.NewCmd(),
+		cleanup.NewCmd(),
 	)
 	return cmd
 }