+++ cmd/operator-sdk/alpha/run/cmd.go
@@ -12,33 +12,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package olm
+package run
 
 import (
-	"fmt"
-	"log"
-
 	olmoperator "github.com/operator-framework/operator-sdk/internal/olm/operator"
 
+	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
-func NewUpCmd() *cobra.Command {
+type runArgs struct {
+	olm bool
+}
+
+func NewCmd() *cobra.Command {
+	cargs := &runArgs{}
 	c := &olmoperator.OLMCmd{}
 	cmd := &cobra.Command{
-		Use:   "up",
-		Short: "Deploy an operator using the Operator Lifecycle Manager",
+		Use:   "run",
+		Short: "Run an Operator in a variety of environments",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) != 1 {
-				return fmt.Errorf("command %q requires exactly one argument", cmd.CommandPath())
-			}
-			c.ManifestsDir = args[0]
-			if err := c.Up(); err != nil {
-				log.Fatalf("Failed to deploy operator: %v", err)
+			switch {
+			case cargs.olm:
+				if err := c.Run(); err != nil {
+					log.Fatalf("Failed to run operator: %v", err)
+				}
 			}
 			return nil
 		},
 	}
+	// OLM is the default.
+	cmd.Flags().BoolVar(&cargs.olm, "olm", true, "The operator to be run will be managed by OLM in a cluster.")
+	// TODO(estroz): refactor flag setting when new run mode options are added.
 	c.AddToFlagSet(cmd.Flags())
 	return cmd
 }